+++ packages/snap/snap.manifest.json
@@ -15,7 +15,7 @@
         "registry": "https://registry.npmjs.org"
       }
     },
-    "shasum": "PqqcsZCtUe1U7HthbkREuVBdimxg+Hoy48Q8QOC+IWI="
+    "shasum": "F2qW8DTisr9quDMpGmAlozVxiKhmUje2Tdz1QH3IG2k="
   },
   "initialPermissions": {
     "endowment:ethereum-provider": {},
